making sure we close http resources
